Trait owned_chunks::StaticOwnedChunks

pub trait StaticOwnedChunks: Sized {
    type Iter<const CHUNK_SZ: usize>;

    // Required method
    fn static_owned_chunks<const CHUNK_SZ: usize>(self) -> Self::Iter<CHUNK_SZ>;
}

A variant of OwnedChunks with const chunk size; this primarily exists to allow storage size optimizations for types where storage size depends on a constant, like arrays. For further information see the docs of array::ArrayChunks.

Example

use owned_chunks::StaticOwnedChunks;

fn take_ownership(v: Vec<i32>) {
    // some implementation
}

for chunk in [vec![1, 2, 3], vec![4, 5, 6], vec![7, 8, 9]].static_owned_chunks::<2>() {
    for v in chunk {
        take_ownership(v);
    }
}

Required Associated Types

type Iter<const CHUNK_SZ: usize>

Required Methods

fn static_owned_chunks<const CHUNK_SZ: usize>(self) -> Self::Iter<CHUNK_SZ>

Object Safety

This trait is not object safe.